+++ singer_sdk/helpers/_compat.py
@@ -12,4 +12,5 @@ except ImportError:
     # Running on pre-3.8 Python; use importlib-metadata package
     import importlib_metadata as metadata  # type: ignore
 
+
 __all__ = ["metadata", "final"]